requirements.txt
# Core dependencies
z3-solver==4.12.2.0
boto3==1.34.0
botocore==1.34.0

# AWS Lambda utilities
aws-lambda-powertools==2.31.0
aws-xray-sdk==2.12.1

# Data processing
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2023.3

# Ontology processing
rdflib==7.0.0
owlready2==0.45

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7
ujson==5.9.0

# Monitoring
prometheus-client==0.19.0